// frontend/libs/shared/data-access/src/lib/generated/models/notification-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NotificationResponse {
  body: string;
  created_at: string;
  id: string;
  is_read: boolean;
  notification_type: string;
  reference_id?: (string | null);
  reference_type?: (string | null);
  tenant_id: string;
  title: string;
}